features/doctors-info: make doctor schedule in appointments service use doctor user id

+++ BackEnd/src/Doctors/Doctors.API/Doctors/Mappers/QueryMapper.cs
@@ -26,7 +26,7 @@ public static class QueryMapper
     public static WorkDayResponse ToResponse(
         this WorkDayDto workDay)
         => new(
-            workDay.DayOfWeek,
+            (int)workDay.DayOfWeek,
             workDay.WorkTimes.Select(s => s.ToResponse()).ToList());
     
     public static WorkTimeRangeResponse ToResponse(
